# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gym Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Branch {
  id: number;
  name: string;
  address: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  capacity?: number | null;
  isActive?: boolean;
  createdAt: string;
}

export interface BranchInput {
  name: string;
  address: string;
  phone: string;
  email?: string;
  capacity?: number;
  isActive?: boolean;
}

export interface BranchUpdate {
  name?: string;
  address?: string;
  phone?: string;
  email?: string;
  capacity?: number;
  isActive?: boolean;
}

export interface Trainer {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  specialization: string;
  /** @nullable */
  bio?: string | null;
  branchId: number;
  isActive?: boolean;
  createdAt: string;
}

export interface TrainerInput {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  specialization: string;
  bio?: string;
  branchId: number;
  isActive?: boolean;
}

export interface TrainerUpdate {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  specialization?: string;
  bio?: string;
  branchId?: number;
  isActive?: boolean;
}

export interface Member {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  isActive?: boolean;
  joinedAt: string;
  createdAt: string;
}

export interface MemberInput {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dateOfBirth?: string;
  address?: string;
  emergencyContact?: string;
  isActive?: boolean;
}

export interface MemberUpdate {
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  dateOfBirth?: string;
  address?: string;
  emergencyContact?: string;
  isActive?: boolean;
}

export interface GymClass {
  id: number;
  name: string;
  category: string;
  /** @nullable */
  description?: string | null;
  durationMinutes: number;
  maxCapacity: number;
  difficultyLevel?: string;
  isActive?: boolean;
  createdAt: string;
}

export interface GymClassInput {
  name: string;
  category: string;
  description?: string;
  durationMinutes: number;
  maxCapacity: number;
  difficultyLevel?: string;
  isActive?: boolean;
}

export interface GymClassUpdate {
  name?: string;
  category?: string;
  description?: string;
  durationMinutes?: number;
  maxCapacity?: number;
  difficultyLevel?: string;
  isActive?: boolean;
}

export interface ClassSchedule {
  id: number;
  classId: number;
  /** @nullable */
  className?: string | null;
  trainerId: number;
  /** @nullable */
  trainerName?: string | null;
  branchId: number;
  /** @nullable */
  branchName?: string | null;
  startTime: string;
  endTime: string;
  status?: string;
  currentBookings: number;
  /** @nullable */
  maxCapacity?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClassScheduleInput {
  classId: number;
  trainerId: number;
  branchId: number;
  startTime: string;
  endTime: string;
  status?: string;
  notes?: string;
}

export interface ClassScheduleUpdate {
  classId?: number;
  trainerId?: number;
  branchId?: number;
  startTime?: string;
  endTime?: string;
  status?: string;
  notes?: string;
}

export interface Booking {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  scheduleId: number;
  /** @nullable */
  className?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface BookingInput {
  memberId: number;
  scheduleId: number;
  notes?: string;
}

export interface BookingUpdate {
  status?: string;
  notes?: string;
}

export interface MembershipPlan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  durationDays: number;
  price: number;
  /** @nullable */
  maxClassesPerMonth?: number | null;
  isActive?: boolean;
  createdAt: string;
}

export interface MembershipPlanInput {
  name: string;
  description?: string;
  durationDays: number;
  price: number;
  maxClassesPerMonth?: number;
  isActive?: boolean;
}

export interface MembershipPlanUpdate {
  name?: string;
  description?: string;
  durationDays?: number;
  price?: number;
  maxClassesPerMonth?: number;
  isActive?: boolean;
}

export interface Membership {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  planId: number;
  /** @nullable */
  planName?: string | null;
  startDate: string;
  endDate: string;
  status: string;
  createdAt: string;
}

export interface MembershipInput {
  memberId: number;
  planId: number;
  startDate: string;
  status?: string;
}

export interface MembershipUpdate {
  planId?: number;
  startDate?: string;
  endDate?: string;
  status?: string;
}

export interface Payment {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  /** @nullable */
  membershipId?: number | null;
  amount: number;
  type: string;
  status: string;
  /** @nullable */
  method?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  memberId: number;
  membershipId?: number;
  amount: number;
  type: string;
  status?: string;
  method?: string;
  notes?: string;
}

export interface PaymentUpdate {
  status?: string;
  method?: string;
  notes?: string;
}

export interface DashboardStats {
  totalMembers: number;
  activeMembers: number;
  totalTrainers: number;
  totalClasses: number;
  scheduledClassesToday: number;
  totalRevenue: number;
  revenueThisMonth: number;
  bookingsToday: number;
  totalBranches?: number;
  activeMemberships?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  entityId?: number | null;
  createdAt: string;
}

export interface ClassUtilization {
  classId: number;
  className: string;
  category?: string;
  totalSessions: number;
  totalBookings: number;
  avgUtilizationPercent: number;
}

export interface RevenueMonth {
  month: string;
  revenue: number;
  paymentCount: number;
}