# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all gym branches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get a branch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "404":
          description: Not found
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Trainers ---
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all gym class types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GymClass"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a gym class type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GymClassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get a gym class type by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
        "404":
          description: Not found
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a gym class type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GymClassUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a gym class type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Schedules ---
...
  /membership-plans:
    get:
      operationId: listMembershipPlans
      tags: [membership-plans]
      summary: List all membership plans
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MembershipPlan"
    post:
      operationId: createMembershipPlan
      tags: [membership-plans]
      summary: Create a membership plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipPlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipPlan"
  /membership-plans/{id}:
    get:
      operationId: getMembershipPlan
      tags: [membership-plans]
      summary: Get a membership plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipPlan"
        "404":
          description: Not found
    patch:
      operationId: updateMembershipPlan
      tags: [membership-plans]
      summary: Update a membership plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipPlanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipPlan"
    delete:
      operationId: deleteMembershipPlan
      tags: [membership-plans]
      summary: Delete a membership plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Memberships ---
  /memberships:
    get:
      operationId: listMemberships
      tags: [memberships]
      summary: List member memberships
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Membership"
    post:
      operationId: createMembership
      tags: [memberships]
      summary: Assign membership to a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
  /memberships/{id}:
    get:
      operationId: getMembership
      tags: [memberships]
      summary: Get a membership by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "404":
          description: Not found
    patch:
      operationId: updateMembership
      tags: [memberships]
      summary: Update a membership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
    delete:
      operationId: deleteMembership
      tags: [memberships]
      summary: Delete a membership
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Payments ---
...
components:
  schemas:
...
    Branch:
      type: object
      required: [id, name, address, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        capacity:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name, address, phone]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        capacity:
          type: integer
        isActive:
          type: boolean
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        capacity:
          type: integer
        isActive:
          type: boolean

    # Trainer
...
    GymClass:
      type: object
      required: [id, name, category, durationMinutes, maxCapacity, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        maxCapacity:
          type: integer
        difficultyLevel:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    GymClassInput:
      type: object
      required: [name, category, durationMinutes, maxCapacity]
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        maxCapacity:
          type: integer
        difficultyLevel:
          type: string
        isActive:
          type: boolean
    GymClassUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        maxCapacity:
          type: integer
        difficultyLevel:
          type: string
        isActive:
          type: boolean

    # ClassSchedule
...
    MembershipPlan:
      type: object
      required: [id, name, durationDays, price, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        durationDays:
          type: integer
        price:
          type: number
        maxClassesPerMonth:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    MembershipPlanInput:
      type: object
      required: [name, durationDays, price]
      properties:
        name:
          type: string
        description:
          type: string
        durationDays:
          type: integer
        price:
          type: number
        maxClassesPerMonth:
          type: integer
        isActive:
          type: boolean
    MembershipPlanUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationDays:
          type: integer
        price:
          type: number
        maxClassesPerMonth:
          type: integer
        isActive:
          type: boolean

    # Membership
    Membership:
      type: object
      required: [id, memberId, planId, startDate, endDate, status, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: ["string", "null"]
        planId:
          type: integer
        planName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
        createdAt:
          type: string
    MembershipInput:
      type: object
      required: [memberId, planId, startDate]
      properties:
        memberId:
          type: integer
        planId:
          type: integer
        startDate:
          type: string
        status:
          type: string
    MembershipUpdate:
      type: object
      properties:
        planId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string

    # Payment
...