# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gym Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  name: string;
  address: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  capacity?: number | null;
  isActive?: boolean;
  createdAt: string;
}