# lib/api-zod/src/generated/types/branchUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gym Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface BranchUpdate {
  name?: string;
  address?: string;
  phone?: string;
  email?: string;
  capacity?: number;
  isActive?: boolean;
}