# lib/api-zod/src/generated/types/membership.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gym Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Membership {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  planId: number;
  /** @nullable */
  planName?: string | null;
  startDate: string;
  endDate: string;
  status: string;
  createdAt: string;
}