# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gym Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipPlan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  durationDays: number;
  price: number;
  /** @nullable */
  maxClassesPerMonth?: number | null;
  isActive?: boolean;
  createdAt: string;
}