# lib/api-zod/src/generated/types/membershipPlanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Gym Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipPlanInput {
  name: string;
  description?: string;
  durationDays: number;
  price: number;
  maxClassesPerMonth?: number;
  isActive?: boolean;
}